+++ core/impl/src/main/java/com/blazebit/persistence/impl/BaseUpdateCriteriaBuilderImpl.java
@@ -89,6 +89,12 @@ public abstract class BaseUpdateCriteriaBuilderImpl<T, X extends BaseUpdateCrite
 
     @Override
     @SuppressWarnings("unchecked")
+    public X setNull(String attributeName) {
+        return setExpression(attributeName, "NULL");
+    }
+
+    @Override
+    @SuppressWarnings("unchecked")
     public X setExpression(String attributeName, String expression) {
         verifyBuilderEnded();
         addAttribute(attributeName);